requirements.txt
amqp==5.1.0
anyio==3.5.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-filter==21.1
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flower==1.0.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
humanize==4.0.0
idna==3.3
kombu==5.2.4
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
